refactor: move account type check to `useAccountType` (#907)
